refactor(qb): SJRA-000 rename upload-id-modal to upload-id-facet
