+++ frontend/components/stories/query-builder/query-builder.stories.tsx
@@ -923,7 +923,7 @@ export const SearchFacet: Story = {
   ),
 };
 
-export const UploadIdModal: Story = {
+export const UploadIdFacet: Story = {
   parameters: {
     msw: {
       handlers: [
